big update

+++ Project_ALTP/app/src/main/java/clc65/quanggck/project_altp/model/Question.java
@@ -5,9 +5,9 @@ public class Question {
     public int id;
     public String content;
 
-    public String a,b,c,d;
+    public String a, b, c, d;
 
-    public float rate_a, rate_b, rate_c, rate_d;
+    public int rate_a, rate_b, rate_c, rate_d;
 
     public String correct;        // A, B, C, D
     public Difficulty difficulty; // LEVEL_1 → LEVEL_15
@@ -16,8 +16,8 @@ public class Question {
     public Question() {
     }
 
-    public Question(int id, String content, String a, String b, String c, String d, float rate_a,
-                    float rate_b, float rate_c, float rate_d, String correct, Difficulty difficulty) {
+    public Question(int id, String content, String a, String b, String c, String d,
+                    int rate_a, int rate_b, int rate_c, int rate_d, String correct, Difficulty difficulty) {
         this.id = id;
         this.content = content;
         this.a = a;
@@ -32,3 +32,5 @@ public class Question {
         this.difficulty = difficulty;
     }
 }
+
+
